+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### Changed (v4.1.2)
+
+## 2019-04-14
+
+### Android
+
+- Added an option to start Android picker in "All Photos" closes [#111](https://github.com/Sh1d0w/multi_image_picker/issues/111)
+- Added selectCircleStrokeColor closes [#113](https://github.com/Sh1d0w/multi_image_picker/issues/113)
+  
 ### Changed (v4.1.1)
 
 ## 2019-04-14